apple_qe_portfolio_final/code/netqual/requirements.txt
pytest>=7.0
pytest-timeout>=2.1
allure-pytest>=2.13
